[CORAL-6] Added utitlity to tests translation of all hive views in metastore.
Added tool to read all the hive views in the configure datastore and
test the translation of those views to Presto sql. This is useful for
testing coverage. The tool is intentionally not integrated with
automated testing because it requires external setup and credentials
of the user to connect to grid using kerberos.

RB=1184556
R=atumbde,khtran
A=khtran