Control casting of view relnode
This change limits scenarios when a cast operations is generated for views
relnodes if the base table or view schema has evolved. See LIHADOOP-34428
for details. If a struct column of base table has evolved to add more fields,
Dali allows these fields to reflect in the view even though these are not part
of the view definition. We try to match this Dali behavior.
1. Do not generate cast operators if the struct columns of base table have evolved
to add new fields. Deletion of fields is not expected.
2. Avoid cast operations between char and varchar types if those only differ by precision.
3. Turn off validation between generated relnode rowtype and validated rowtype. This is required
to support above two scenarios.

Testing: Integration tests on production views

RB=1228424
R=wmoustaf,khtran
A=khtran