[CORAL-47] - Add rewrite logic of GenericProject placeholder query to be Spark specific
Add rewrite logic to transform a GenericProject placeholder in coral-hive to a Spark specific UDF call.
Included in this RB is logic that creates a Spark DataType JSON String for a desired schema.
This JSON string schema is parseable by Spark DataType and a Spark-native GenericProject UDF handles the transformations.

RB=1603186
BUG=CORAL-47,LIHADOOP-44834
G=dali-reviewers
R=rratti,wmoustaf,atumbde,smahadik,khtran
A=rratti,atumbde